+++ crates/sevenmark-server/src/connection/mod.rs
@@ -1,2 +1,2 @@
 pub mod database_conn;
-pub mod seaweedfs_conn;
+pub mod r2_conn;